requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.1
pydantic==2.5.3